Update openapi to accept any content for artifact uploads

+++ common/src/main/java/io/apicurio/registry/rest/v2/SearchResource.java
@@ -30,7 +30,7 @@ public interface SearchResource {
       @QueryParam("order") SortOrder order, @QueryParam("orderby") SortBy orderby,
       @QueryParam("labels") List<String> labels, @QueryParam("properties") List<String> properties,
       @QueryParam("description") String description, @QueryParam("group") String group,
-      @QueryParam("globalId") Long globalId, @QueryParam("contentId") Long contentId);
+      @QueryParam("globalId") int globalId, @QueryParam("contentId") int contentId);
 
   /**
    * Returns a paginated list of all artifacts with at least one version that matches the
